refactor: extract read-only domains behind service facade
