+++ src/diptrace_mcp/services/__init__.py
@@ -1,0 +1,2 @@
+"""Internal domain services composed by :class:`DipTraceService`."""
+
